# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/packages": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "Get all packages",
                "operationId": "d09b9ef6ed46b5c4f8e680afbc23b8e3",
                "responses": {
                    "200": {
                        "description": "Packages retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Packages"
                ],
                "summary": "Create a new package",
                "operationId": "c393ac414f125f39ed06f3abeeb4308e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "SenderID",
                                    "CustomerID",
                                    "weight",
                                    "package_type",
                                    "shipping_cost",
                                    "receipt_date"
                                ],
                                "properties": {
                                    "SenderID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "CustomerID": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "package_type": {
                                        "type": "string",
                                        "example": "Box"
                                    },
                                    "shipping_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000
                                    },
                                    "receipt_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Package created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/packages/{id}": {
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "Update a package",
                "operationId": "31121c1278670481ee6e92b8ae1c7497",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "SenderID": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "CustomerID": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "package_type": {
                                        "type": "string",
                                        "example": "Envelope"
                                    },
                                    "shipping_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000
                                    },
                                    "receipt_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-27"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Package updated successfully"
                    },
                    "404": {
                        "description": "Package not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "Delete a package",
                "operationId": "38718113909921c59043663af143d1ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package deleted successfully"
                    },
                    "404": {
                        "description": "Package not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}